MetadataV14 : unit test and cleanup

+++ polkaj-scale-types/src/main/java/io/emeraldpay/polkaj/scaletypes/v14/MetadataContainer.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Data
 public class MetadataContainer {
-    private Integer magicNumber;
+    private long magicNumber;
     private Integer version;
     private Metadata metadata;
